feat(admin): group Sync Providers and Custom LLMs under Gateway tabs
Consolidate the standalone Sync Providers and Custom LLMs admin pages
into a single Gateway page with tab navigation, following the same
pattern used by the Alerting section. The old URLs redirect to the
new tabbed page.
